PanelDue, UI: reduced colourscheme and alert dependencies
