+++ src/UI/UserInterface.hpp
@@ -16,6 +16,8 @@
 #include <UI/ColourSchemes.hpp>
 #include <UI/Display.hpp>
 #include <UI/Events.hpp>
+#include <General/String.h>
+#include <General/StringFunctions.h>
 
 extern IntegerField *freeMem;
 extern StaticTextField *debugField;
@@ -23,12 +25,44 @@ extern StaticTextField *touchCalibInstruction;
 extern StaticTextField *messageTextFields[], *messageTimeFields[];
 extern TextField *fwVersionField;
 
-class Alert;
+const size_t alertTextLength = 165;			// maximum characters in the alert text
+const size_t alertTitleLength = 50;			// maximum characters in the alert title
+
+struct Alert
+{
+	int32_t mode;
+	uint32_t seq;
+	uint32_t controls;
+	float timeout;
+	String<50> title;
+	String<alertTextLength> text;
+	Bitmap<uint8_t> flags;
+
+	static constexpr uint8_t GotMode = 0;
+	static constexpr uint8_t GotSeq = 1;
+	static constexpr uint8_t GotTimeout = 2;
+	static constexpr uint8_t GotTitle = 3;
+	static constexpr uint8_t GotText = 4;
+	static constexpr uint8_t GotControls = 5;
+	static constexpr uint8_t GotAll =
+			(1 << GotMode)
+			| (1 << GotSeq)
+			| (1 << GotTimeout)
+			| (1 << GotTitle)
+			| (1 << GotText)
+			| (1 << GotControls);
+
+	Alert() : mode(0), seq(0), controls(0), timeout(0.0) { flags.Clear(); }
+
+	bool AllFlagsSet() const { return flags.GetRaw() == GotAll; }
+};
+
 
 namespace UI
 {
 	extern unsigned int GetNumLanguages();
 	extern void CreateFields(uint32_t language, const ColourScheme& colours, uint32_t p_infoTimeout);
+	extern void InitColourScheme(const ColourScheme *scheme);
 	extern void ActivateScreensaver();
 	extern void DeactivateScreensaver();
 	extern void AnimateScreensaver();
